UNET/website/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
segmentation-models-pytorch
opencv-python
numpy
pillow
